feat: 添加差异分析功能及工具注册
- 在 types.ts 中新增 DiffAnalysis 接口，用于描述代码差异分析结果，包括新增行、删除行和上下文行的信息。
- 更新 ai-code-review/helper.ts 文件，新增 analyzeDiffLines 函数，解析 diff 字符串并提取行号信息。
- 注册 AI 代码审查工具，提供分析合并请求变更和推送代码审查评论的功能。
- 在 file-operation 和 merge-request 模块中注册相关工具，增强文件操作和合并请求管理功能。
