+++ src/server/tools/ai-code-review/index.ts
@@ -1,121 +1,13 @@
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import {  z } from "zod";
-import { GitLabClient } from "../../gitlab/client";
+import { GitLabClient } from "../../../gitlab/client";
 import type { 
-  AICodeReviewResult, 
   CodeReviewIssue, 
-  CodeReviewRules, 
-  FileFilterConfig,
-  CodeReviewReport,
-  GitLabFileChange 
-} from "../../config/types";
+    } from "../../../config/types";
+import { analyzeDiffLines, filterReviewableChanges, getCodeReviewRules, getExclusionReason, getFileExtension, shouldReviewFile } from "./helper";
 
-/**
- * 代码审查规则配置
- */
-const CODE_REVIEW_RULES: Record<string, CodeReviewRules> = {
-  '.ts': {
-    focus_areas: ['类型安全', '性能优化', '代码规范', '错误处理'],
-    specific_rules: [
-      '确保所有函数都有明确的返回类型',
-      '避免使用 any 类型，使用具体类型或泛型',
-      '异步函数必须正确处理错误',
-      '组件和函数名使用 PascalCase 或 camelCase',
-      '避免在循环中进行异步操作',
-      '使用 const assertions 来获得更精确的类型推断'
-    ],
-    ignore_patterns: ['*.d.ts', '*.test.ts', '*.spec.ts'],
-    severity_mapping: {
-      'any类型使用': 'warning',
-      '未处理的Promise': 'critical',
-      '循环中的异步操作': 'warning',
-      '缺少类型注解': 'suggestion'
-    }
-  },
-  '.vue': {
-    focus_areas: ['组件设计', '性能优化', 'Vue最佳实践', '可访问性'],
-    specific_rules: [
-      'template 中避免复杂的表达式',
-      '合理使用 computed 和 watch',
-      '确保组件的 props 有正确的类型定义',
-      'emit 事件需要明确定义类型',
-      '避免在 template 中直接修改 props',
-      '使用正确的生命周期钩子'
-    ],
-    ignore_patterns: [],
-    severity_mapping: {
-      'template表达式过于复杂': 'warning',
-      'props缺少类型': 'warning',
-      '直接修改props': 'critical'
-    }
-  },
-  '.js': {
-    focus_areas: ['ES6+语法', '性能', '错误处理', '代码规范'],
-    specific_rules: [
-      '使用 const/let 替代 var',
-      '使用箭头函数或 function 声明',
-      '避免回调地狱，使用 Promise 或 async/await',
-      '正确处理异常和错误',
-      '使用严格模式',
-      '避免全局变量污染'
-    ],
-    ignore_patterns: ['*.min.js', '*.bundle.js'],
-    severity_mapping: {
-      '使用var声明': 'suggestion',
-      '回调地狱': 'warning',
-      '未处理异常': 'critical'
-    }
-  }
-};
 
-/**
- * 文件过滤配置
- */
-const FILE_FILTER_CONFIG: FileFilterConfig = {
-  includedExtensions: ['.ts', '.tsx', '.js', '.jsx', '.vue', '.py', '.java', '.cs', '.go', '.rs'],
-  excludePatterns: [
-    /node_modules/,
-    /dist/,
-    /build/,
-    /\.git/,
-    /coverage/,
-    /\.nyc_output/,
-    /\.next/,
-    /\.nuxt/,
-    /\.vuepress/,
-    /\.temp/,
-    /\.cache/,
-    /\.DS_Store/,
-    /\.env/,
-    /\.log$/,
-    /\.lock$/,
-    /package-lock\.json$/,
-    /yarn\.lock$/,
-    /pnpm-lock\.yaml$/,
-    /\.map$/,
-    /\.min\./,
-    /\.d\.ts$/,
-    /\.test\./,
-    /\.spec\./,
-    /\.stories\./,
-    /\.config\./,
-    /\.md$/,
-    /README/i,
-    /CHANGELOG/i,
-    /LICENSE/i
-  ],
-  maxFileSize: 100 * 1024, // 100KB
-  maxDiffLines: 500
-};
 
-/**
- * 差异分析结果
- */
-interface DiffAnalysis {
-  newLines: Array<{lineNumber: number, content: string}>;
-  deletedLines: Array<{lineNumber: number, content: string}>;
-  contextLines: Array<{lineNumber: number, content: string}>;
-}
 
 /**
  * 评论分级策略
@@ -583,145 +475,3 @@ export function registerAICodeReviewTools(server: McpServer, gitlabClient: GitLa
     }
   );
 }
-
-/**
- * 分析diff并提取准确的行号信息
- */
-function analyzeDiffLines(diff: string): DiffAnalysis {
-  const lines = diff.split('\n');
-  const newLines = [];
-  const deletedLines = [];
-  const contextLines = [];
-  
-  let newLineNumber = 0;
-  let oldLineNumber = 0;
-  
-  for (const line of lines) {
-    if (line.startsWith('@@')) {
-      // 解析行号范围 @@-oldStart,oldCount +newStart,newCount@@
-      const match = line.match(/@@ -(\d+),?\d* \+(\d+),?\d* @@/);
-      if (match) {
-        oldLineNumber = parseInt(match[1]) - 1;
-        newLineNumber = parseInt(match[2]) - 1;
-      }
-    } else if (line.startsWith('+') && !line.startsWith('+++')) {
-      newLineNumber++;
-      newLines.push({ lineNumber: newLineNumber, content: line.substring(1) });
-    } else if (line.startsWith('-') && !line.startsWith('---')) {
-      oldLineNumber++;
-      deletedLines.push({ lineNumber: oldLineNumber, content: line.substring(1) });
-    } else if (line.startsWith(' ')) {
-      oldLineNumber++;
-      newLineNumber++;
-      contextLines.push({ lineNumber: newLineNumber, content: line.substring(1) });
-    }
-  }
-  
-  return { newLines, deletedLines, contextLines };
-}
-
-/**
- * 获取文件扩展名
- */
-function getFileExtension(filePath: string): string {
-  const match = filePath.match(/\.([^.]+)$/);
-  return match ? `.${match[1]}` : '';
-}
-
-/**
- * 获取文件的代码审查规则
- */
-function getCodeReviewRules(extension: string): CodeReviewRules {
-  const rules = CODE_REVIEW_RULES[extension];
-  if (!rules) {
-    return {
-      focus_areas: ['基础代码质量', '可读性', '维护性'],
-      specific_rules: [
-        '确保代码逻辑清晰',
-        '避免过度复杂的嵌套',
-        '函数和变量命名要有意义',
-        '添加必要的注释'
-      ],
-      ignore_patterns: [],
-      severity_mapping: {}
-    };
-  }
-  
-  return rules;
-}
-
-/**
- * 判断文件是否需要审查
- */
-function shouldReviewFile(filePath: string, extension: string): boolean {
-  // 检查扩展名
-  if (!FILE_FILTER_CONFIG.includedExtensions.includes(extension)) {
-    return false;
-  }
-  
-  // 检查排除模式
-  for (const pattern of FILE_FILTER_CONFIG.excludePatterns) {
-    if (pattern.test(filePath)) {
-      return false;
-    }
-  }
-  
-  return true;
-}
-
-/**
- * 过滤可审查的变更
- */
-function filterReviewableChanges(changes: GitLabFileChange[], focusFiles?: string[]): GitLabFileChange[] {
-  return changes.filter(change => {
-    // 如果指定了重点文件，只审查指定文件
-    if (focusFiles && focusFiles.length > 0) {
-      return focusFiles.some(file => change.new_path.includes(file));
-    }
-    
-    // 排除删除的文件
-    if (change.deleted_file) {
-      return false;
-    }
-    
-    // 检查文件是否应该被审查
-    const extension = getFileExtension(change.new_path);
-    if (!shouldReviewFile(change.new_path, extension)) {
-      return false;
-    }
-    
-    // 检查diff大小
-    if (change.diff && change.diff.split('\n').length > FILE_FILTER_CONFIG.maxDiffLines) {
-      console.log(`⚠️  文件 ${change.new_path} diff过大，跳过审查`);
-      return false;
-    }
-    
-    return true;
-  });
-}
-
-/**
- * 获取文件被排除的原因
- */
-function getExclusionReason(change: GitLabFileChange): string {
-  if (change.deleted_file) {
-    return '文件已删除';
-  }
-  
-  const extension = getFileExtension(change.new_path);
-  if (!FILE_FILTER_CONFIG.includedExtensions.includes(extension)) {
-    return '文件类型不支持';
-  }
-  
-  for (const pattern of FILE_FILTER_CONFIG.excludePatterns) {
-    if (pattern.test(change.new_path)) {
-      return '匹配排除模式';
-    }
-  }
-  
-  if (change.diff && change.diff.split('\n').length > FILE_FILTER_CONFIG.maxDiffLines) {
-    return 'diff过大';
-  }
-  
-  return '未知原因';
-}
