+++ src/server/tools/index.ts
@@ -1,4 +1,4 @@
 // 导出所有工具注册函数
-export { registerMergeRequestTools } from "./merge-request.js";
-export { registerFileOperationTools } from "./file-operations.js";
-export { registerAICodeReviewTools } from "./ai-code-review.js"; 
+export { registerMergeRequestTools } from "./merge-request/index.js";
+export { registerFileOperationTools } from "./file-operation/index.js";
+export { registerAICodeReviewTools } from "./ai-code-review/index.js";
